Rename `signals.rs` to `signal.rs`
This matches the naming of the upstream crates.
